Linted varch.

+++ arch/varch.sh
@@ -9,21 +9,21 @@ TEST_ARCHIVE=false
 _remove_lock=true
 
 on_sigterm() {
-    echo "Caught user interrupt, closing down."
+    printf 'Caught user interrupt, closing down.\n'
     if [[ -f $LOCK ]] && $_remove_lock; then
-        rm -f $LOCK
+        rm -f "$LOCK"
     else
-        echo "Warning: can't remove lock file."
+        printf 'Warning: can'\''t remove lock file.'
     fi
     exit 1
 }
 
 on_exit() {
-    echo "Finished."
+    printf 'Finished.\n'
     if [[ -f $LOCK ]] && $_remove_lock; then
-        rm -f $LOCK
+        rm -f "$LOCK"
     else
-        echo "Warning: can't remove lock file."
+        printf 'Warning: can'\''t remove lock file.\n'
     fi
     exit 0
 }
@@ -32,10 +32,9 @@ trap on_sigterm SIGTERM SIGINT
 trap on_exit EXIT
 
 read_password() {
-    printf "Password: "
-    read -s password
+    read -rsp 'Password: ' password
     printf "\n"
-    if [[ $(echo $password | sha256sum) != $(cat $PASSWD) ]]; then
+    if [[ $(echo "$password" | sha256sum) != $(cat "$PASSWD") ]]; then
         echo "Incorrect password."
         exit 1
     fi
@@ -43,18 +42,18 @@ read_password() {
 
 # Lock check
 if [[ -f $LOCK ]]; then
-    echo >&2 'This script is already running.'
+    printf >&2 'This script is already running.\n'
     _remove_lock=false
-    read
+    read -r
     exit 1
 fi
 
 # Sanity tests
 if [[ ! -d $DEST ]]; then
-    echo >&2 'Destination directory does not exist.'
+    printf >&2 'Destination directory does not exist.\n'
     exit 1
 elif [[ ! -f $PASSWD ]]; then
-    echo >&2 'Cannot find archive password.'
+    printf >&2 'Cannot find archive password.\n'
     exit 1
 fi
 
@@ -65,21 +64,22 @@ varch() {
     touch "$LOCK"
 
     if [[ -d $1 ]]; then
-        echo "[Compressing]"
+        printf '[Compressing]\n'
         read_password
-        printf "Hashing images...\n"
+        printf 'Hashing images...\n'
         $HASH_SCRIPT "$DEST/$1"
-        printf "Backing up old archive...\n"
-        [ -f $ARCHIVE ] && cp -u $ARCHIVE $ARCHIVE~
-        printf "Adding files to archive...\n"
-        7z a -p"$password" -t7z -m0=lzma -mx=8 -mhe=on $ARCHIVE $1
-        $TEST_ARCHIVE && 7z t -p"$password" $ARCHIVE
-        printf "Removing old files...\n"
-        rm -r $1/
+        [[ -d $1/.git ]] && git_update "$1"
+        printf 'Backing up old archive...\n'
+        [[ -f $ARCHIVE ]] && mv -u "$ARCHIVE" "$ARCHIVE~"
+        printf 'Adding files to archive...\n'
+        7z a -p"$password" -t7z -ms=on -mhe=on -m0=lzma -mx=3 "$ARCHIVE" "$1"
+        $TEST_ARCHIVE && 7z t -p"$password" "$ARCHIVE"
+        printf 'Removing old files...\n'
+        rm -r "${1:?}/"
         if $CLEAR_RECENT; then
-            printf "Clearing recent documents...\n"
+            printf 'Clearing recent documents...\n'
             : > ~/.local/share/recently-used.xbel
-            printf "Clearing thumbnails...\n"
+            printf 'Clearing thumbnails...\n'
 
             if [[ -d ~/.thumbnails ]]; then
                 rm -f ~/.thumbnails/normal/*
@@ -89,15 +89,18 @@ varch() {
                 rm -f ~/.cache/thumbnails/large/*
             fi
         fi
+    elif [[ ! -f $ARCHIVE ]]; then
+        printf '[Error]\n'
+        printf 'Cannot find archive at %s!\n' "$ARCHIVE"
     else
-        echo "[Extracting]"
+        printf '[Extracting]\n'
         read_password
-        7z x -p"$password" $ARCHIVE
+        7z x -p"$password" "$ARCHIVE"
     fi
 }
 
 if [[ $# -eq 0 ]]; then
-    read -p "Which archive would you like to access? " name
+    read -rp 'Which archive would you like to access? ' name
     varch "$name"
 elif [[ $# -gt 0 ]]; then
     varch "$1"
